Move nkbg tests to seperate file (error 1128 for windows builds).
