+++ networkit/cpp/io/test/CMakeLists.txt
@@ -1,7 +1,8 @@
 networkit_add_test(io IOGTest
         auxiliary community generators structures)
 
+networkit_add_test(io NetworkitBinaryGTest)
+
 networkit_add_test(io MemoryMappedFileGTest)
 
 networkit_add_benchmark(io IOBenchmark geometric)
-
